Fix: Improper error handling in ball encoding.
Ball encoder was catching Key errors that often occurred because undefined symbols where referenced in a vol3 json.

This however also caught key errors due to erros in the ball encoder.

There now is a dedicated exception for attempting to encode a type that is not defined in in the symbols.