Update: Encoding unknown structs
Pointers to "unencodable" types are now represented in the symbol graph since they will be encoded in the memory graphs as well.

This mostly refers to void and function pointers.

These typedescriptors will be reprsented with block type vectors completely made up of 0s.

To make this more explicit, a new blocktype "unknown" was introduced, which is also used for paddings and "unidentifyable" parts of structs like unions.