Ex: Dense Follows with sampling.
An experiment: We encode multiple "precedes" and "follows" nodes in symbol graphs using the alternative implementation of _compute_precdes. Then, in any training epoch, we use dgls sampling facilities to train with only one "preceding" and "following" nodes of the multiple that are possible.

Intuition: Null pointers will not get a chunk in memory encoding,
thus we often get a different "preceding" and "following" node in the symbol graphs and the memory graphs.